Revert "Fake table.c.values() type: List[Column]"
Using .values() is unsafe, as a new column added to roardata
will implicitly get referenced by calls to .values() even if
roardata hasn't otherwise been changed to use the column.

This means that safely deploying a migration to a table used
in roardata is very inconvenient, unless it can be guaranteed
to go out in an all-deploy.

This reverts commit fb0936343d2efd5d3f35e4a0e93557a232be9426.